backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
llama-cpp-python
Pillow==10.2.0
python-dateutil==2.8.2
aiofiles==23.2.1
